services/txt-to-json/requirements.txt
aiofiles==24.1.0
redis==5.2.1
aio-pika==9.5.4
minio==7.2.15
google-generativeai==0.8.4
